--- a/lib/codeintel/lsif/conversion/datastructures/default_idset_map.go
+++ b/lib/codeintel/lsif/conversion/datastructures/default_idset_map.go
@@ -1,25 +1,33 @@
 package datastructures
 
-// DefaultIDSetMap is a space-efficient map from integer keys to identifier sets. This
-// map adds convenience operations that will operate on the set at a specific key, and
-// create it if it does not yet exist in the map.
+const (
+	// random sentinel key to identify runtime errors
+	uninit_sentinel_key = -0xc0c0
+)
+
+// DefaultIDSetMap is a small-size-optimized map from integer keys to identifier sets.
+// It adds convenience operations that operate on the set for a specific key.
 //
 // The correlation process creates many such maps (e.g. result set and contains relations),
-// the majority of which contain only a single element. This structure optimizes for the
-// case where only a single key is present in the map.
+// the majority of which contain only a single element. Since Go maps have high overhead
+// (see https://golang.org/src/runtime/map.go#L115), we optimize for the common case.
 //
 // For concrete numbers, processing an index for aws-sdk-go produces 1.5 million singleton
 // maps, and only 25k non-singleton maps.
 //
-// Each map starts out empty. Insertion of the first element sets a key and value field
-// directly in the map struct. Insertion of a second element will promote the struct into
-// a non-singleton map and heap-allocate a builtin map to hold additional keys. Maps have
-// large overhead (see https://golang.org/src/runtime/map.go#L115), so we only want to pay
-// this cost when we need to.
+// The map is conceptually in one of three states:
+// - Empty: This is the initial state.
+// - Inline: This contains an inline element.
+// - Heap: This contains key-value pairs in a Go map.
+//
+// The state of the map may change:
+// - On additions: Empty → Inline or Inline → Heap.
+// - On deletions: Inline → Empty or Heap → Inline.
 type DefaultIDSetMap struct {
-	key   int            // singleton map key
-	value *IDSet         // singleton map value
-	m     map[int]*IDSet // non-singleton map
+	len         int            // number of keys
+	inlineKey   int            // key for the Inline state
+	inlineValue *IDSet         // value for the Inline state
+	m           map[int]*IDSet // storage for 2 or more key-value pairs
 }
 
 // NewDefaultIDSetMap creates a new empty default identifier set map.
@@ -28,19 +36,46 @@ func NewDefaultIDSetMap() *DefaultIDSetMap {
 }
 
 // DefaultIDSetMapWith creates a default identifier set map with the given contents.
+//
+// The supplied map is not used directly if it has length 0 or 1.
 func DefaultIDSetMapWith(m map[int]*IDSet) *DefaultIDSetMap {
-	return &DefaultIDSetMap{m: m}
+	switch len(m) {
+	case 0:
+		return NewDefaultIDSetMap()
+	case 1:
+		tmp := NewDefaultIDSetMap()
+		for k, v := range m {
+			tmp.inlineKey = k
+			tmp.inlineValue = v
+		}
+		tmp.len = 1
+		return tmp
+	default:
+		return &DefaultIDSetMap{len: len(m), m: m}
+	}
+}
+
+// Len returns the number of keys.
+func (sm *DefaultIDSetMap) Len() int {
+	if sm == nil {
+		return 0
+	}
+	return sm.len
 }
 
 // Get returns the identifier set at the given key or nil if it does not exist.
 func (sm *DefaultIDSetMap) Get(key int) *IDSet {
-	if sm.key == key {
-		return sm.value
-	}
-	if sm.m != nil {
+	switch sm.Len() {
+	case 0:
+		return nil
+	case 1:
+		if sm.inlineKey == key {
+			return sm.inlineValue
+		}
+		return nil
+	default:
 		return sm.m[key]
 	}
-	return nil
 }
 
 // Pop returns the identifier set at the given key or nil if it does not exist and
@@ -64,75 +99,103 @@ func (sm *DefaultIDSetMap) Pop(key int) *IDSet {
 
 // Delete removes the identifier set at the given key if it exists.
 func (sm *DefaultIDSetMap) Delete(key int) {
-	if sm.key == key {
-		sm.key = 0
-		sm.value = nil
-	}
-	if sm.m != nil {
+	switch sm.Len() {
+	case 0:
+		return
+	case 1:
+		if sm.inlineKey == key {
+			sm.inlineKey = uninit_sentinel_key
+			sm.inlineValue = nil
+			sm.len = 0
+		}
+	default:
+		if _, ok := sm.m[key]; ok {
+			sm.len--
+		}
 		delete(sm.m, key)
+		if sm.len == 1 {
+			for k, v := range sm.m {
+				sm.inlineKey = k
+				sm.inlineValue = v
+			}
+			sm.m = nil
+		}
 	}
 }
 
 // Each invokes the given function with each key and identifier set in the map.
+//
+// The order of iteration is not guaranteed to be deterministic.
 func (sm *DefaultIDSetMap) Each(f func(key int, value *IDSet)) {
-	if sm.key != 0 {
-		f(sm.key, sm.value)
-	}
-	if sm.m != nil {
+	switch sm.len {
+	case 0:
+		return
+	case 1:
+		f(sm.inlineKey, sm.inlineValue)
+	default:
 		for k, v := range sm.m {
 			f(k, v)
 		}
 	}
 }
 
-// SetLen returns the number of identifiers in the identifier set at the given key.
-func (sm *DefaultIDSetMap) SetLen(key int) int {
-	if sm.key == key {
-		return sm.value.Len()
-	}
-	if sm.m != nil {
+// NumIDsForKey returns the number of identifiers in the identifier set at the given key.
+func (sm *DefaultIDSetMap) NumIDsForKey(key int) int {
+	switch sm.len {
+	case 0:
+		return 0
+	case 1:
+		if sm.inlineKey == key {
+			return sm.inlineValue.Len()
+		}
+	default:
 		if s, ok := sm.m[key]; ok {
 			return s.Len()
 		}
 	}
-
 	return 0
 }
 
-// SetContains determines if the given identifier belongs to the set at the given key.
-func (sm *DefaultIDSetMap) SetContains(key, id int) bool {
-	if sm.key == key {
-		return sm.value.Contains(id)
-	}
-	if sm.m != nil {
+// Contains determines if the given identifier belongs to the set at the given key.
+func (sm *DefaultIDSetMap) Contains(key, id int) bool {
+	switch sm.len {
+	case 0:
+		return false
+	case 1:
+		return sm.inlineKey == key && sm.inlineValue.Contains(id)
+	default:
 		if s, ok := sm.m[key]; ok {
 			return s.Contains(id)
 		}
 	}
-
 	return false
 }
 
-// SetEach invokes the given function with each identifier in the set at the given key.
-func (sm *DefaultIDSetMap) SetEach(key int, f func(id int)) {
-	if sm.key == key {
-		sm.value.Each(f)
+// EachID invokes the given function with each identifier in the set at the given key.
+//
+// The order of iteration is not guaranteed to be deterministic.
+func (sm *DefaultIDSetMap) EachID(key int, f func(id int)) {
+	switch sm.len {
+	case 0:
 		return
-	}
-	if sm.m != nil {
+	case 1:
+		if sm.inlineKey == key {
+			sm.inlineValue.Each(f)
+		}
+	default:
 		if s, ok := sm.m[key]; ok {
 			s.Each(f)
 		}
 	}
 }
 
-// SetAdd inserts an identifier into the set at the given key.
-func (sm *DefaultIDSetMap) SetAdd(key, id int) {
+// AddID inserts an identifier into the set at the given key.
+func (sm *DefaultIDSetMap) AddID(key, id int) {
 	sm.getOrCreate(key).Add(id)
 }
 
-// SetUnion inserts all the identifiers of other into the set a the given key.
-func (sm *DefaultIDSetMap) SetUnion(key int, other *IDSet) {
+// UnionIDSet inserts all the identifiers of other into the set a the given key.
+func (sm *DefaultIDSetMap) UnionIDSet(key int, other *IDSet) {
 	if other == nil || other.Len() == 0 {
 		return
 	}
@@ -140,41 +203,35 @@ func (sm *DefaultIDSetMap) SetUnion(key int, other *IDSet) {
 	sm.getOrCreate(key).Union(other)
 }
 
-// GetOrCreate will return the set at the given key, or create an empty set if it does not exist.
+// getOrCreate will return the set at the given inlineKey, or create an empty set if it does not exist.
+//
+// The return value is never nil.
 func (sm *DefaultIDSetMap) getOrCreate(key int) *IDSet {
-	if sm.key == key {
-		return sm.value
-	}
-	if sm.m != nil {
+	switch sm.len {
+	case 0:
+		sm.inlineKey = key
+		sm.inlineValue = NewIDSet()
+		sm.len = 1
+		return sm.inlineValue
+	case 1:
+		if sm.inlineKey == key {
+			return sm.inlineValue
+		}
+		newValue := NewIDSet()
+		sm.m = map[int]*IDSet{sm.inlineKey: sm.inlineValue, key: newValue}
+		sm.len = 2
+		sm.inlineValue = nil
+		sm.inlineKey = uninit_sentinel_key
+		return newValue
+	default:
 		if s, ok := sm.m[key]; ok {
 			return s
 		}
+		newValue := NewIDSet()
+		sm.m[key] = newValue
+		sm.len++
+		return newValue
 	}
-
-	if sm.m != nil {
-		// Already a map, just add a new key
-		s := NewIDSet()
-		sm.m[key] = s
-		return s
-	}
-
-	if sm.key == 0 {
-		// Populating a singleton set
-		s := NewIDSet()
-		sm.key = key
-		sm.value = s
-		return s
-	}
-
-	// Adding to a singleton set. Create a new map and add the
-	// new value along with the old singleton value to the it
-	s := NewIDSet()
-	sm.m = make(map[int]*IDSet, 2)
-	sm.m[key] = s
-	sm.m[sm.key] = sm.value
-	sm.key = 0
-	sm.value = nil
-	return s
 }
 
 // compareDefaultIDSetMaps returns true if the given identifier default identifier set maps
@@ -184,16 +241,12 @@ func compareDefaultIDSetMaps(x, y *DefaultIDSetMap) bool {
 		return true
 	}
 
-	if x == nil || y == nil {
+	if x.len != y.len {
 		return false
 	}
 
 	m1 := toMap(x)
 	m2 := toMap(y)
-
-	if len(m1) != len(m2) {
-		return false
-	}
 
 	for k, v := range m1 {
 		if !compareIDSets(v, m2[k]) {
@@ -207,19 +260,16 @@ func compareDefaultIDSetMaps(x, y *DefaultIDSetMap) bool {
 // toMap returns a copy of the map backing the default identifier set map. This is called from
 // compareDefaultIDSetMaps for testing and should not be used in the hot path.
 func toMap(s *DefaultIDSetMap) map[int]*IDSet {
-	if s.key != 0 {
-		return map[int]*IDSet{
-			s.key: s.value,
-		}
-	}
-
-	if s.m != nil {
+	switch s.len {
+	case 0:
+		return nil
+	case 1:
+		return map[int]*IDSet{s.inlineKey: s.inlineValue}
+	default:
 		m := map[int]*IDSet{}
 		for k, v := range s.m {
 			m[k] = v
 		}
 		return m
 	}
-
-	return nil
 }