--- a/lib/codeintel/lsif/conversion/datastructures/default_idset_map.go
+++ b/lib/codeintel/lsif/conversion/datastructures/default_idset_map.go
@@ -1,8 +1,12 @@
 package datastructures
 
 const (
+	mapEmpty         = 0
+	mapInline        = 1
+	mapHeap          = 2
+	ILLEGAL_MAPSTATE = "invariant violation: illegal map state!"
 	// random sentinel key to identify runtime errors
-	uninit_sentinel_key = -0xc0c0
+	uninitSentinelKey = -0xc0c0
 )
 
 // DefaultIDSetMap is a small-size-optimized map from integer keys to identifier sets.
@@ -24,7 +28,6 @@ const (
 // - On additions: Empty → Inline or Inline → Heap.
 // - On deletions: Inline → Empty or Heap → Inline.
 type DefaultIDSetMap struct {
-	len         int            // number of keys
 	inlineKey   int            // key for the Inline state
 	inlineValue *IDSet         // value for the Inline state
 	m           map[int]*IDSet // storage for 2 or more key-value pairs
@@ -35,91 +38,108 @@ func NewDefaultIDSetMap() *DefaultIDSetMap {
 	return &DefaultIDSetMap{}
 }
 
-// DefaultIDSetMapWith creates a default identifier set map with the given contents.
-//
-// The supplied map is not used directly if it has length 0 or 1.
-func DefaultIDSetMapWith(m map[int]*IDSet) *DefaultIDSetMap {
-	switch len(m) {
-	case 0:
-		return NewDefaultIDSetMap()
-	case 1:
-		tmp := NewDefaultIDSetMap()
-		for k, v := range m {
-			tmp.inlineKey = k
-			tmp.inlineValue = v
+func (sm *DefaultIDSetMap) state() uint8 {
+	if sm.inlineValue == nil {
+		if sm.m == nil {
+			return mapEmpty
 		}
-		tmp.len = 1
-		return tmp
-	default:
-		return &DefaultIDSetMap{len: len(m), m: m}
+		return mapHeap
 	}
+	if sm.m != nil {
+		panic("m field of DefaultIDSetMap should be nil when value is present inline")
+	}
+	return mapInline
+}
+
+// DefaultIDSetMapWith creates a default identifier set map with
+// a copy of the given contents.
+//
+// map entries with nil or empty IDSets are ignored.
+func DefaultIDSetMapWith(m map[int]*IDSet) *DefaultIDSetMap {
+	tmp := NewDefaultIDSetMap()
+	for k, v := range m {
+		tmp.UnionIDSet(k, v)
+	}
+	return tmp
 }
 
 // Len returns the number of keys.
 func (sm *DefaultIDSetMap) Len() int {
-	if sm == nil {
+	switch sm.state() {
+	case mapEmpty:
 		return 0
+	case mapInline:
+		return 1
+	case mapHeap:
+		return len(sm.m)
+	default:
+		panic(ILLEGAL_MAPSTATE)
 	}
-	return sm.len
 }
 
 // Get returns the identifier set at the given key or nil if it does not exist.
 func (sm *DefaultIDSetMap) Get(key int) *IDSet {
-	switch sm.Len() {
-	case 0:
+	switch sm.state() {
+	case mapEmpty:
 		return nil
-	case 1:
+	case mapInline:
 		if sm.inlineKey == key {
 			return sm.inlineValue
 		}
 		return nil
-	default:
+	case mapHeap:
 		return sm.m[key]
+	default:
+		panic(ILLEGAL_MAPSTATE)
 	}
 }
 
 // Pop returns the identifier set at the given key or nil if it does not exist and
 // removes the key from the map.
 func (sm *DefaultIDSetMap) Pop(key int) *IDSet {
-	if sm.key == key {
-		v := sm.value
-		sm.key = 0
-		sm.value = nil
+	switch sm.state() {
+	case mapEmpty:
+		return nil
+	case mapInline:
+		if sm.inlineKey != key {
+			return nil
+		}
+		v := sm.inlineValue
+		sm.inlineKey = uninitSentinelKey
+		sm.inlineValue = nil
 		return v
-	}
-	if sm.m != nil {
+	case mapHeap:
 		v, ok := sm.m[key]
 		if ok {
 			delete(sm.m, key)
 		}
 		return v
+	default:
+		panic(ILLEGAL_MAPSTATE)
 	}
-	return nil
 }
 
 // Delete removes the identifier set at the given key if it exists.
 func (sm *DefaultIDSetMap) Delete(key int) {
-	switch sm.Len() {
-	case 0:
+	switch sm.state() {
+	case mapEmpty:
 		return
-	case 1:
+	case mapInline:
 		if sm.inlineKey == key {
-			sm.inlineKey = uninit_sentinel_key
+			sm.inlineKey = uninitSentinelKey
 			sm.inlineValue = nil
-			sm.len = 0
 		}
-	default:
-		if _, ok := sm.m[key]; ok {
-			sm.len--
-		}
+	case mapHeap:
 		delete(sm.m, key)
-		if sm.len == 1 {
+		if len(sm.m) == 1 {
 			for k, v := range sm.m {
 				sm.inlineKey = k
 				sm.inlineValue = v
 			}
 			sm.m = nil
 		}
+	default:
+		panic(ILLEGAL_MAPSTATE)
 	}
 }
 
@@ -127,46 +147,52 @@ func (sm *DefaultIDSetMap) Delete(key int) {
 //
 // The order of iteration is not guaranteed to be deterministic.
 func (sm *DefaultIDSetMap) Each(f func(key int, value *IDSet)) {
-	switch sm.len {
-	case 0:
+	switch sm.state() {
+	case mapEmpty:
 		return
-	case 1:
+	case mapInline:
 		f(sm.inlineKey, sm.inlineValue)
-	default:
+	case mapHeap:
 		for k, v := range sm.m {
 			f(k, v)
 		}
+	default:
+		panic(ILLEGAL_MAPSTATE)
 	}
 }
 
 // NumIDsForKey returns the number of identifiers in the identifier set at the given key.
 func (sm *DefaultIDSetMap) NumIDsForKey(key int) int {
-	switch sm.len {
-	case 0:
+	switch sm.state() {
+	case mapEmpty:
 		return 0
-	case 1:
+	case mapInline:
 		if sm.inlineKey == key {
 			return sm.inlineValue.Len()
 		}
-	default:
+	case mapHeap:
 		if s, ok := sm.m[key]; ok {
 			return s.Len()
 		}
+	default:
+		panic(ILLEGAL_MAPSTATE)
 	}
 	return 0
 }
 
 // Contains determines if the given identifier belongs to the set at the given key.
 func (sm *DefaultIDSetMap) Contains(key, id int) bool {
-	switch sm.len {
-	case 0:
+	switch sm.state() {
+	case mapEmpty:
 		return false
-	case 1:
+	case mapInline:
 		return sm.inlineKey == key && sm.inlineValue.Contains(id)
-	default:
+	case mapHeap:
 		if s, ok := sm.m[key]; ok {
 			return s.Contains(id)
 		}
+	default:
+		panic(ILLEGAL_MAPSTATE)
 	}
 	return false
 }
@@ -175,17 +201,19 @@ func (sm *DefaultIDSetMap) Contains(key, id int) bool {
 //
 // The order of iteration is not guaranteed to be deterministic.
 func (sm *DefaultIDSetMap) EachID(key int, f func(id int)) {
-	switch sm.len {
-	case 0:
+	switch sm.state() {
+	case mapEmpty:
 		return
-	case 1:
+	case mapInline:
 		if sm.inlineKey == key {
 			sm.inlineValue.Each(f)
 		}
-	default:
+	case mapHeap:
 		if s, ok := sm.m[key]; ok {
 			s.Each(f)
 		}
+	default:
+		panic(ILLEGAL_MAPSTATE)
 	}
 }
 
@@ -207,30 +235,29 @@ func (sm *DefaultIDSetMap) UnionIDSet(key int, other *IDSet) {
 //
 // The return value is never nil.
 func (sm *DefaultIDSetMap) getOrCreate(key int) *IDSet {
-	switch sm.len {
-	case 0:
+	switch sm.state() {
+	case mapEmpty:
 		sm.inlineKey = key
 		sm.inlineValue = NewIDSet()
-		sm.len = 1
 		return sm.inlineValue
-	case 1:
+	case mapInline:
 		if sm.inlineKey == key {
 			return sm.inlineValue
 		}
 		newValue := NewIDSet()
 		sm.m = map[int]*IDSet{sm.inlineKey: sm.inlineValue, key: newValue}
-		sm.len = 2
 		sm.inlineValue = nil
-		sm.inlineKey = uninit_sentinel_key
+		sm.inlineKey = uninitSentinelKey
 		return newValue
-	default:
+	case mapHeap:
 		if s, ok := sm.m[key]; ok {
 			return s
 		}
 		newValue := NewIDSet()
 		sm.m[key] = newValue
-		sm.len++
 		return newValue
+	default:
+		panic(ILLEGAL_MAPSTATE)
 	}
 }
 
@@ -241,7 +268,7 @@ func compareDefaultIDSetMaps(x, y *DefaultIDSetMap) bool {
 		return true
 	}
 
-	if x.len != y.len {
+	if x.state() != y.state() {
 		return false
 	}
 
@@ -260,16 +287,18 @@ func compareDefaultIDSetMaps(x, y *DefaultIDSetMap) bool {
 // toMap returns a copy of the map backing the default identifier set map. This is called from
 // compareDefaultIDSetMaps for testing and should not be used in the hot path.
 func toMap(s *DefaultIDSetMap) map[int]*IDSet {
-	switch s.len {
-	case 0:
+	switch s.state() {
+	case mapEmpty:
 		return nil
-	case 1:
+	case mapInline:
 		return map[int]*IDSet{s.inlineKey: s.inlineValue}
-	default:
+	case mapHeap:
 		m := map[int]*IDSet{}
 		for k, v := range s.m {
 			m[k] = v
 		}
 		return m
+	default:
+		panic(ILLEGAL_MAPSTATE)
 	}
 }