@varungandhi-src 2022-02-11
I did two runs for a couple of different patches (the before and after ones for `client/web`). Looks like the change is quite small, and we should get a much bigger speedup after implementing Chris's suggestion:

```
before this PR:
  - large patch: 2m27s, 2m23s
  - small patch: 28s, 30s
after this PR:
  - large patch: 2m36s, 2m33s
  - small patch: 32s, 31s
```